# app_nodos/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Tree Management API",
        "version": "1.0.0",
        "description": "API para gestionar árboles de nodos con soporte multi-idioma"
    },
    "paths": {
...
        "/api/users/{id}/nodes-created/": {
            "get": {
                "operationId": "users_nodes_created_retrieve",
                "description": "Endpoint custom para listar los nodos CREADOS por un usuario específico.\nPermite a ADMIN/SUDO ver todos; USER solo puede ver los suyos.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "A unique integer value identifying this user.",
                        "required": true
                    }
                ],
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/me/": {
            "get": {
                "operationId": "users_me_retrieve",
                "description": "Endpoint para recuperar el perfil del usuario autenticado (GET /me).",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/me/change-password/": {
            "post": {
                "operationId": "users_me_change_password_create",
                "description": "Endpoint para que el usuario actual cambie su propia contraseña.",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/me/update/": {
            "put": {
                "operationId": "users_me_update_update",
                "description": "Endpoint para actualizar el perfil del usuario autenticado.",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "patch": {
                "operationId": "users_me_update_partial_update",
                "description": "Endpoint para actualizar el perfil del usuario autenticado.",
                "tags": [
                    "users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUser"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUser"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedUser"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PatchedUser": {
                "type": "object",
                "description": "Serializer base para operaciones generales de lectura/actualización de usuarios.\n\nCampos clave:\n- 'password': WRITE_ONLY, requerido solo en creación/actualización de contraseña.\n- Campos de auditoría de nodos ('nodes_created_count' en DetailSerializer)\n  se exponen como READ_ONLY.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "username": {
                        "type": "string",
                        "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
                        "pattern": "^[\\w.@+-]+$",
                        "maxLength": 150
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Correo electrónico único del usuario.",
                        "maxLength": 254
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "role": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RoleEnum"
                            }
                        ],
                        "description": "Rol de acceso del usuario (SUDO, ADMIN, USER).\n\n* `SUDO` - Super User Ops\n* `ADMIN` - Administrador\n* `USER` - Usuario Regular"
                    },
                    "is_email_confirmed": {
                        "type": "boolean",
                        "description": "Indica si el usuario ha verificado su correo electrónico."
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Active",
                        "description": "Designates whether this user should be treated as active. Unselect this instead of deleting accounts."
                    },
                    "is_staff": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Staff status",
                        "description": "Designates whether the user can log into this admin site."
                    },
                    "is_superuser": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Superuser status",
                        "description": "Designates that this user has all permissions without explicitly assigning them."
                    },
                    "date_joined": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "last_login": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "description": "Contraseña. Se ignora en el listado, requerida en el POST/PUT."
                    }
                }
            },
            "RoleEnum": {
                "enum": [
                    "SUDO",
                    "ADMIN",
                    "USER"
                ],
                "type": "string",
                "description": "* `SUDO` - Super User Ops\n* `ADMIN` - Administrador\n* `USER` - Usuario Regular"
            },
...
            "User": {
                "type": "object",
                "description": "Serializer base para operaciones generales de lectura/actualización de usuarios.\n\nCampos clave:\n- 'password': WRITE_ONLY, requerido solo en creación/actualización de contraseña.\n- Campos de auditoría de nodos ('nodes_created_count' en DetailSerializer)\n  se exponen como READ_ONLY.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "username": {
                        "type": "string",
                        "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
                        "pattern": "^[\\w.@+-]+$",
                        "maxLength": 150
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Correo electrónico único del usuario.",
                        "maxLength": 254
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "role": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RoleEnum"
                            }
                        ],
                        "description": "Rol de acceso del usuario (SUDO, ADMIN, USER).\n\n* `SUDO` - Super User Ops\n* `ADMIN` - Administrador\n* `USER` - Usuario Regular"
                    },
                    "is_email_confirmed": {
                        "type": "boolean",
                        "description": "Indica si el usuario ha verificado su correo electrónico."
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Active",
                        "description": "Designates whether this user should be treated as active. Unselect this instead of deleting accounts."
                    },
                    "is_staff": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Staff status",
                        "description": "Designates whether the user can log into this admin site."
                    },
                    "is_superuser": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Superuser status",
                        "description": "Designates that this user has all permissions without explicitly assigning them."
                    },
                    "date_joined": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "last_login": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "description": "Contraseña. Se ignora en el listado, requerida en el POST/PUT."
                    }
                },
                "required": [
                    "date_joined",
                    "email",
                    "id",
                    "is_staff",
                    "is_superuser",
                    "last_login",
                    "username"
                ]
            },
...
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}